feat(money): store money as integer cents and discounts in basis points
Replace floating point money in the backend with integer cents and
discounts as basis points, migrating existing databases and backups in
a staging area before activation. Add real-flow, storage-format and IPC
test suites, a single-instance storage lock, and update the frontend to
convert directly between BRL text and cents.

+++ src-tauri/src/models/inventory_movement.rs
@@ -12,7 +12,7 @@ pub struct InventoryMovement {
     pub reference_os_id: Option<String>,
     pub os_display_id: Option<String>,
     pub reason: String,
-    pub unit_cost: Option<f64>,
+    pub unit_cost: Option<i64>,
     pub created_at: Option<String>,
 }
